Docs updated for line join, cap, & miter limit features
