+++ packages/graphics/src/styles/LineStyle.ts
@@ -1,5 +1,5 @@
 import { FillStyle } from './FillStyle';
-import { LINE_JOIN, LINE_CAP } from '@pixi/constants';
+import { LINE_JOIN, LINE_CAP } from '../const';
 
 /**
  * Represents the line style for Graphics.
