+++ packages/graphics/src/utils/buildLine.ts
@@ -3,7 +3,7 @@ import { Point, SHAPES } from '@pixi/math';
 import type { Polygon } from '@pixi/math';
 import type { GraphicsData } from '../GraphicsData';
 import type { GraphicsGeometry } from '../GraphicsGeometry';
-import { LINE_JOIN, LINE_CAP } from '@pixi/constants';
+import { LINE_JOIN, LINE_CAP } from '../const';
 import { GRAPHICS_CURVES } from '../const';
 
 /**
